회원이 입력한 비밀번호가 db에 들어있는 회원의 비밀번호와 일치할 때 모든 기록,정보를 지우도록 설계.

+++ src/main/java/com/example/accessingdatamysql/User/UserService.java
@@ -89,15 +89,20 @@ public class UserService {
 
     public String deleteUser(String token, String rawPassword){
         String email = jwtUtil.extractEmail(token);
-        Integer userId = userRepository.findByEmail(email).getId();
 
-        userRepository.deleteById(userId);
-        userStartRecordRepository.deleteAllByUserId(userId);
-        userCessationRecordRepository.deleteAllByUserId(userId);
-        missionRepository.deleteAllByUserid(userId);
-        missionRecordRepository.deleteAllByUserId(userId);
+        if(authenticate(token, rawPassword)){
+            Integer userId = userRepository.findByEmail(email).getId();
 
-        return "deleted";
+            userRepository.deleteById(userId);
+            userStartRecordRepository.deleteAllByUserId(userId);
+            userCessationRecordRepository.deleteAllByUserId(userId);
+            missionRepository.deleteAllByUserid(userId);
+            missionRecordRepository.deleteAllByUserId(userId);
+
+            return "deleted";
+        }else{
+            return "password does not match";
+        }
     }
 
 }
